Trying to make deathScreen work (BROKEN ATM)
DO NOT MERGE THIS WITH MAIN
So, deathScreen classes have been modelled after MainMenuScreen properly this time, they have similar actions, it should be all set up mostly and MainGameScreen actually calls the correct enum when the player dies. I just can't work out how to make it show when the character dies. Currently this is a broken version, pushed so others can work out why it crashes. it crashes when the character dies with the issue being the serviceLocator can't load the image, which seems wrong??
**I REPEAT DO NOT MERGE THIS WITH MAIN**
-Elias