Death buttons & Win Screen
**Win Screen:**
Currently due to enemies not existing in level 2 and the lack of level 2 plug coordinates there is no Win condition. To win the game press 'N'.

**Death Screen**
Death screen's for level 1 and 2 have transparent buttons in the correct place that take you to the correct screens. In true rougelike fashion if you press play again you start the game again entirely with no save.

Sound has been added to win and death screen also to better suit the theme and mood attempting to be conveyed in those states. (Shout out to Isaac for coaching me through transparent buttons)